requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
boto3==1.34.0
cryptography==41.0.7
urllib3==2.1.0
email-validator==2.1.0
